% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f.R
\name{f}
\alias{f}
\title{TEST
A little test to see if you are able}
\usage{
f(x, y)
}
\arguments{
\item{x}{this is the indipendent variable}

\item{y}{this is the dependent variable}
}
\description{
TEST
A little test to see if you are able
}
