% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function1.R
\name{function1}
\alias{function1}
\title{Title of your function}
\usage{
function1(arg1, arg2)
}
\arguments{
\item{arg1}{Description of the first argument.}

\item{arg2}{Description of the second argument.}
}
\value{
Description of what the function returns.
}
\description{
A short description of what your function does.
}
