% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.R
\docType{data}
\name{input}
\alias{input}
\title{Dane wejsciowe dla funkcji \code{fun_pack()}}
\format{
Zawartosc danych:
\describe{
\item{Nat}{Natezenie ruchu}
\item{Segment}{Segment pojazdow}
\item{Fuel}{Rodzaj paliwa}
\item{Technology}{Zastosowana technologia pracy silnika}
}
}
\usage{
input
}
\description{
Dane wejsciowe dla funkcji \code{fun_pack()}
}
\examples{
\dontrun{
input
}
}
\keyword{datasets}
