% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvnpdf_invC.R
\name{mvnpdf_invC}
\alias{mvnpdf_invC}
\title{Multivariate Gaussian Distribution}
\usage{
mvnpdf_invC(x, mean = rep(0, nrow(x)), varcovM = diag(nrow(x)), Log = TRUE)
}
\arguments{
\item{x}{a matrix with n columns and p rows}

\item{mean}{a vector of means}

\item{Log}{logical parameter, with default value to TRUE.}

\item{varcoM}{a variance-covariance matrix}
}
\value{
return a list containing the matrix x, and a vector of length n of the multivariate normal distribution density values at those points
}
\description{
Description
}
\details{
Details
}
\examples{
y <- mvpdf(x = matrix(c(1,2)) ,mean=rep(1,2),varcoM = diag(2),Log=TRUE)
}
