% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvpdfoptim_para.R
\name{mvnpdfoptim_para}
\alias{mvnpdfoptim_para}
\title{Multivariate Gaussian Distribution}
\usage{
mvnpdfoptim_para(
  x,
  mean = rep(0, nrow(x)),
  varcovM = diag(nrow(x)),
  Log = TRUE
)
}
\arguments{
\item{x}{a matrix with n columns and p rows}

\item{mean}{a vector of means}

\item{Log}{logical parameter, with default value to TRUE.}

\item{varcoM}{a variance-covariance matrix}
}
\value{
return a list containing the matrix x, and a vector of length n of the multivariate normal distribution density values at those points
}
\description{
Description
}
\details{
Details
}
