% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvpdf.R
\name{mvpdf}
\alias{mvpdf}
\title{Multivariate Gaussian Distribution}
\usage{
mvpdf(x, mean = rep(0, nrow(x)), varcoM = diag(nrow(x)), Log = TRUE)
}
\arguments{
\item{x}{a matrix with n columns and p rows}

\item{mean}{a vector of means}

\item{varcoM}{a variance-covariance matrix}

\item{Log}{logical parameter, with default value to TRUE.}
}
\value{
return a list containing the matrix x, and a vector of length n of the multivariate normal distribution density values at those points
}
\description{
Description
}
\details{
Details
}
\examples{
y <- mvpdf(x = matrix(c(1,2)) ,mean=rep(1,2),varcoM = diag(2),Log=TRUE)
}
