% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myci.R
\name{myci}
\alias{myci}
\title{Confidence Interval Function}
\usage{
myci(x)
}
\arguments{
\item{x}{vector of values a random sample}
}
\value{
A 95% confidence interval for a given sample/vector, 'x'.
}
\description{
Confidence Interval Function
}
\examples{

}
