% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myhyper.R
\name{myhyper}
\alias{myhyper}
\title{Hypergeometric Function}
\usage{
myhyper(iter = 100, N = 20, r = 12, n = 5)
}
\arguments{
\item{n}{}
}
\value{
a plot of the hypergeometric probability distribution using given parameters.
}
\description{
Generates a plot of a hypergeometric probability distribution with given data.
}
\examples{
plot100= myhyper(iter=100,n=5, N=20,r=12)

}
