% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{mynewcurve
Lab 6 Function}
\usage{
myncurve(mu, sigma, a)
}
\arguments{
\item{mu}{mean}

\item{a}{}
}
\value{

}
\description{
mynewcurve
Lab 6 Function
}
\examples{

plot(table(rpois(100, 5)), type = "h", col = "red", lwd = 10,
main = "rpois(100, lambda = 5)")
myncurve(mu = 5, sigma = 3, a = 0.05)
}
