% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mytc.R
\name{mytc}
\alias{mytc}
\title{TCalc Function}
\usage{
mytc(x1)
}
\arguments{
\item{x}{vector of a random sample}
}
\value{
The value of the T-Statistic evaluated at the null mean value from a given sample.
}
\description{
TCalc Function
}
\examples{
x1 = rnorm(35, mean = 10, sd =1 )
mytc(x1)


}
