% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NewPack-package.R
\docType{package}
\name{NewPack-package}
\alias{NewPack}
\alias{NewPack-package}
\title{NewPack: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/MounaAbed/NewPack}
  \item \url{https://mounaabed.github.io/NewPack/}
  \item Report bugs at \url{https://github.com/MounaAbed/NewPack/issues}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
\keyword{internal}
