% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfwykres.R
\name{pfwykres}
\alias{pfwykres}
\title{Funkcja rysuje wykres na podstawie obliczonych emisji z funkcji fun_pack()}
\usage{
pfwykres(
  path = out,
  kat = Category,
  default = "Passenger Cars",
  x = Technology,
  y = Emisja
)
}
\arguments{
\item{path}{dataframe}

\item{kat}{character}

\item{default}{character}

\item{x}{character}

\item{y}{character}
}
\value{
dataframe
}
\description{
Funkcja rysuje wykres na podstawie obliczonych emisji z funkcji fun_pack()
}
\examples{
pfwykres(path = out, kat = Category, x = Technology, y = Emisja)

}
