% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mvpdf.R
\name{plot.mvpdf}
\alias{plot.mvpdf}
\title{Plot of the mvpdf function}
\usage{
\method{plot}{mvpdf}(x, ...)
}
\arguments{
\item{x}{an object of class \code{mvpdf} resulting from a call of the
\code{mvpdf()} function.}

\item{...}{graphical parameters passed to \code{plot()} function.}
}
\value{
Nothing is returned, only a plot is given.
}
\description{
Plot of the mvpdf function
}
\examples{
pdfvalues <- mvpdf(x=matrix(seq(-3, 3, by = 0.1), nrow = 1), Log=FALSE)
plot(pdfvalues)
}
