% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wskazniki}
\alias{wskazniki}
\title{Wskazniki emisji z transportu drogowego EEA}
\format{
An object of class \code{data.frame} with 25766 rows and 17 columns.
}
\usage{
wskazniki
}
\description{
Wskazniki emisji z transportu drogowego EEA
}
\examples{
\dontrun{
 wskazniki
 input
}
}
\keyword{datasets}
