% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_pack.R
\name{fun_pack}
\alias{fun_pack}
\title{Funkcja oblicz srednia z losowych zmiennych}
\usage{
fun_pack(a = 10, b = 20, d = 2)
}
\arguments{
\item{a}{double}

\item{b}{double}

\item{d}{double}
}
\value{
double
}
\description{
Funkcja oblicz srednia z losowych zmiennych
}
\examples{
fun_pack(a = 1, b = 500, d = 23)
}
