% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_pack.R
\name{fun_pack}
\alias{fun_pack}
\title{Fukcja obliczania emisji spalania}
\usage{
fun_pack(
  dane = input,
  kategoria = NULL,
  paliwo = NULL,
  euro = NULL,
  technologia = NULL,
  zanieczyszczenie = NULL,
  mod = ""
)
}
\arguments{
\item{dane}{dataframe}

\item{kategoria}{character}

\item{paliwo}{= character}

\item{euro}{character}

\item{technologia}{character}

\item{zanieczyszczenie}{character}

\item{mod}{character}
}
\value{
dataframe
}
\description{
Fukcja obliczania emisji spalania
}
\examples{
fun_pack(input, "Passenger Cars", NULL, "Euro 5", NULL, c("EC", "CO"),"")
}
