% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_pack.R
\name{fun_pack}
\alias{fun_pack}
\title{Test}
\usage{
fun_pack(
  dane = input,
  kategoria = "Passenger Cars",
  paliwo = "Petrol",
  segment = "Mini",
  euro = "Euro 5",
  technologia = "GDI",
  mode = "",
  substancja = c("EC", "CO")
)
}
\arguments{
\item{dane}{dataframe}

\item{kategoria}{character}

\item{euro}{character}

\item{mode}{character}

\item{substancja}{character}
}
\value{

}
\description{
Test
}
