% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NewPack-package.R
\docType{package}
\name{NewPack-package}
\alias{NewPack}
\alias{NewPack-package}
\title{NewPack: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
\keyword{internal}
