% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfwykres.R
\name{pfwykres}
\alias{pfwykres}
\title{Funkcja rysujaca wykres na podstawie obliczonych emisji spalin}
\usage{
pfwykres(
  path = out,
  kat = Category,
  default = "Passenger Cars",
  x = Technology,
  y = Emisja
)
}
\arguments{
\item{path}{dataframe - ramka danych obliczonych emisji z funkcji \code{fun_pack()}}

\item{kat}{character - kategoria pojazdu}

\item{default}{character - domyslna wartosc kategori pojazdu}

\item{x}{character - zmiennna numeryczna (os x)}

\item{y}{character - zmianna kategoryczna (os y)}
}
\value{
double
}
\description{
Funkcja rysuje wykres na podstawie obliczonych emisji z funkcji \code{fun_pack()}
}
\examples{
# Wykres emisji spalin ze wzgledu na technologie pojazdow
pfwykres(path = out,
        kat = Category,
        x = Technology,
        y = Emisja)

}
