% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R, R/wskazniki.R
\docType{data}
\name{wskazniki}
\alias{wskazniki}
\title{Wskazniki emisji z transportu drogowego EEA}
\format{
An object of class \code{data.frame} with 25766 rows and 17 columns.

Zawartosc danych:
\describe{
\item{Category}{Kategoria pojazdu - character}
\item{Fuel}{Rodzaj paliwa - character}
\item{Segment}{Segment pojazdow - character}
\item{Euro.Standard}{Dopuszczalnych norma emisji spalin - character}
\item{Technology}{Technologia pracy silnika - character}
\item{Pollutant}{Rodzaj zanieczyszczenia - character}
\item{Mode}{Tryb jazdy pojazdu - character}
\item{Alpha}{parametr Alpha - double}
\item{Beta}{parametr Beta - double}
\item{Gamma}{parametr Gamma  - double}
\item{Delta}{parametr Delta - double}
\item{Epsilon}{parametr Epsilon - double}
\item{Zita}{parametr Zita - double}
\item{Hta}{parametr Hta - double}
\item{Reduction}{parametr Reduction - double}
\item{Bio}{parametr Bio - double}
\item{Procent}{parametr Procent - double}
}
}
\usage{
wskazniki

wskazniki
}
\description{
Dane zawieraja informacje o emisjach pojazdow roznych kategorii.
}
\examples{
\dontrun{
 wskazniki
 input
}
\dontrun{
 wskazniki
 }
}
\keyword{datasets}
